requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
bidict==0.22.1
black==23.1.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
dill==0.3.6
dnspython==2.3.0
email-validator==1.3.1
eventlet==0.33.3
Faker==16.8.1
fastapi==0.90.1
fastapi-socketio==0.0.9
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.24.2
orjson==3.8.6
packaging==23.0
pandas==1.5.3
pathspec==0.11.0
platformdirs==3.0.0
pydantic==1.10.4
python-dateutil==2.8.2
python-dotenv==0.21.1
python-engineio==4.3.4
python-multipart==0.0.5
python-socketio==5.7.2
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
starlette==0.23.1
tomli==2.0.1
typing_extensions==4.4.0
ujson==5.7.0
urllib3==1.26.14
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websocket-client==1.5.1
websockets==10.4
yarl==1.8.2
